Add DNA related files
Uploaded the needed files to perform nonlinear fitting of the absorbance vs temperature data from the melt-temp DNA data to determine the melting temperature and its uncertainty. A second analysis notebook is also included that performs the nonlinear fitting for the modified van't Hoff equation in order to extract enthalpy and entropy of the reaction, with uncertainties.